requirements.txt
alpaca-py>=0.25.0
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
pandas-market-calendars>=4.3.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
